cm3020-wip/pyproject.toml
[tool.poetry]
name = "cm3020-artificial-intelligence"
version = "0.1.0"
description = ""
authors = ["Mehmet Sunkur <mehmetsunkur@gmail.com>"]
readme = "README.md"

packages = [{include = "parta"}]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
docling = "^2.14.0"
langchain-community = "^0.3.13"
langgraph = "^0.2.60"
langchain-anthropic = "^0.3.1"
langchain-experimental = "^0.3.4"
langchain-openai = "^0.2.14"
pytest = "^8.3.4"
ratelimit = "^2.2.1"
pybullet = "^3.2.6"
duckduckgo-search = "^7.1.1"
scholarly = "^1.7.11"
beautifulsoup4 = "^4.12.3"
pandas = "^2.2.3"
requests = "^2.32.3"
arxiv = "^2.1.3"
tqdm = "^4.67.1"
#unstructured = "^0.16.11"
numpy = "<2.0.0"
markdown = "^3.7"
langchain-ollama = "^0.2.2"
unstructured={version = "^0.15.9", extras = ["doc","docx","pdf","ppt","pptx","odt","md","csv"]}
unstructured_ingest="0.0.8"
onnxruntime-gpu = "^1.20.1"
tensorflow = "^2.18.0"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"


#[[tool.poetry.source]]
#name = "proxpi"
#url = "http://192.168.1.131:9191/index/"
#priority = "primary"


[[tool.poetry.source]]
name = "pytorch-cpu-src"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
